Fix split package for Spring Boot modules
Move io.github.belgif.rest.problem.spring.[client|server] to
io.github.belgif.rest.problem.spring.boot.[client|server] in
Spring Boot modules (split package with belgif-rest-problem-spring).
Marked as a breaking change, but it is unlikely to affect any users,
as there is usually no direct dependency on these classes.

Fixes #325.
