+++ belgif-rest-problem-spring-boot-3-client-starter/src/main/java/io/github/belgif/rest/problem/spring/boot/client/ProblemRestClientCustomizer.java
@@ -1,0 +1,24 @@
+package io.github.belgif.rest.problem.spring.boot.client;
+
+import org.springframework.boot.web.client.RestClientCustomizer;
+import org.springframework.web.client.RestClient;
+
+import io.github.belgif.rest.problem.spring.client.ProblemResponseErrorHandler;
+
+/**
+ * RestClientCustomizer that registers the {@link ProblemResponseErrorHandler}.
+ *
+ * @see ProblemResponseErrorHandler
+ */
+public class ProblemRestClientCustomizer implements RestClientCustomizer {
+
+    private final ProblemResponseErrorHandler errorHandler;
+
+    protected ProblemRestClientCustomizer(ProblemResponseErrorHandler errorHandler) {
+        this.errorHandler = errorHandler;
+    }
+
+    public void customize(RestClient.Builder restClientBuilder) {
+        restClientBuilder.defaultStatusHandler(errorHandler);
+    }
+}
