+++ belgif-rest-problem-spring-boot-3-client-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-io.github.belgif.rest.problem.spring.client.ClientProblemAutoConfiguration
+io.github.belgif.rest.problem.spring.boot.client.ClientProblemAutoConfiguration
