+++ belgif-rest-problem-spring-boot-4-client-starter/src/main/java/io/github/belgif/rest/problem/spring/boot/client/ClientProblemAutoConfiguration.java
@@ -1,0 +1,57 @@
+package io.github.belgif.rest.problem.spring.boot.client;
+
+import org.springframework.boot.autoconfigure.AutoConfiguration;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
+import org.springframework.boot.jackson.autoconfigure.JacksonAutoConfiguration;
+import org.springframework.boot.restclient.RestClientCustomizer;
+import org.springframework.boot.restclient.RestTemplateCustomizer;
+import org.springframework.boot.webclient.WebClientCustomizer;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Import;
+import org.springframework.web.client.RestClient;
+import org.springframework.web.client.RestTemplate;
+import org.springframework.web.reactive.function.client.WebClient;
+
+import io.github.belgif.rest.problem.spring.client.ProblemResponseErrorHandler;
+import io.github.belgif.rest.problem.spring.client.ProblemResponseJackson3ErrorHandler;
+import tools.jackson.databind.ObjectMapper;
+
+/**
+ * Spring Boot AutoConfiguration for rest-problem-spring.
+ */
+@AutoConfiguration
+@Import({ io.github.belgif.rest.problem.spring.ProblemJackson3Configuration.class, JacksonAutoConfiguration.class })
+public class ClientProblemAutoConfiguration {
+
+    @Bean
+    public ProblemResponseErrorHandler problemResponseErrorHandler(ObjectMapper objectMapper) {
+        return new ProblemResponseJackson3ErrorHandler(objectMapper);
+    }
+
+    @ConditionalOnClass({ RestClient.class, RestClientCustomizer.class })
+    public static class RestClientProblemConfiguration {
+        @Bean
+        public ProblemRestClientCustomizer
+                problemRestClientCustomizer(ProblemResponseErrorHandler problemResponseErrorHandler) {
+            return new ProblemRestClientCustomizer(problemResponseErrorHandler);
+        }
+    }
+
+    @ConditionalOnClass({ RestTemplate.class, RestTemplateCustomizer.class })
+    public static class RestTemplateProblemConfiguration {
+        @Bean
+        public ProblemRestTemplateCustomizer
+                problemRestTemplateCustomizer(ProblemResponseErrorHandler problemResponseErrorHandler) {
+            return new ProblemRestTemplateCustomizer(problemResponseErrorHandler);
+        }
+    }
+
+    @ConditionalOnClass({ WebClient.class, WebClientCustomizer.class })
+    public static class WebClientProblemConfiguration {
+        @Bean
+        public ProblemWebClientCustomizer problemWebClientCustomizer() {
+            return new ProblemWebClientCustomizer();
+        }
+    }
+
+}
