+++ belgif-rest-problem-spring-boot-4-client-starter/src/main/java/io/github/belgif/rest/problem/spring/boot/client/ProblemWebClientCustomizer.java
@@ -1,0 +1,17 @@
+package io.github.belgif.rest.problem.spring.boot.client;
+
+import org.springframework.boot.webclient.WebClientCustomizer;
+import org.springframework.web.reactive.function.client.WebClient;
+
+import io.github.belgif.rest.problem.spring.client.WebClientFilter;
+
+/**
+ * WebClientCustomizer that registers a filter that converts problem responses to Problem exceptions.
+ */
+public class ProblemWebClientCustomizer implements WebClientCustomizer {
+
+    public void customize(WebClient.Builder webClientBuilder) {
+        webClientBuilder.filter(WebClientFilter.PROBLEM_FILTER);
+    }
+
+}
