--- belgif-rest-problem-spring-boot-4-client-starter/src/main/java/io/github/belgif/rest/problem/spring/client/ProblemRestClientCustomizer.java
@@ -1,23 +1,0 @@
-package io.github.belgif.rest.problem.spring.client;
-
-import org.springframework.boot.restclient.RestClientCustomizer;
-import org.springframework.web.client.RestClient;
-
-/**
- * RestClientCustomizer that registers the {@link ProblemResponseErrorHandler}.
- *
- * @see ProblemResponseErrorHandler
- */
-public class ProblemRestClientCustomizer implements RestClientCustomizer {
-
-    private final ProblemResponseErrorHandler errorHandler;
-
-    protected ProblemRestClientCustomizer(ProblemResponseErrorHandler errorHandler) {
-        this.errorHandler = errorHandler;
-    }
-
-    public void customize(RestClient.Builder restClientBuilder) {
-        restClientBuilder.defaultStatusHandler(errorHandler);
-    }
-
-}
