+++ belgif-rest-problem-spring-boot-4-client-starter/src/test/java/io/github/belgif/rest/problem/spring/boot/client/ProblemRestClientCustomizerTest.java
@@ -1,0 +1,28 @@
+package io.github.belgif.rest.problem.spring.boot.client;
+
+import static org.assertj.core.api.Assertions.*;
+import static org.mockito.Mockito.*;
+
+import java.util.List;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.test.util.ReflectionTestUtils;
+import org.springframework.web.client.RestClient;
+
+import io.github.belgif.rest.problem.spring.client.ProblemResponseErrorHandler;
+
+class ProblemRestClientCustomizerTest {
+
+    @Test
+    void customize() {
+        ProblemResponseErrorHandler handler = mock(ProblemResponseErrorHandler.class);
+        ProblemRestClientCustomizer customizer = new ProblemRestClientCustomizer(handler) {
+        };
+        RestClient.Builder builder = RestClient.builder();
+        customizer.customize(builder);
+
+        List<?> statusHandlers = (List<?>) ReflectionTestUtils.getField(builder, "statusHandlers");
+        assertThat(statusHandlers).hasSize(1);
+    }
+
+}
