+++ belgif-rest-problem-spring-boot-4-client-starter/src/test/java/io/github/belgif/rest/problem/spring/boot/client/ProblemRestTemplateCustomizerTest.java
@@ -1,0 +1,24 @@
+package io.github.belgif.rest.problem.spring.boot.client;
+
+import static org.assertj.core.api.Assertions.*;
+import static org.mockito.Mockito.*;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.web.client.RestTemplate;
+
+import io.github.belgif.rest.problem.spring.client.ProblemResponseErrorHandler;
+
+class ProblemRestTemplateCustomizerTest {
+
+    @Test
+    void customize() {
+        ProblemResponseErrorHandler handler = mock(ProblemResponseErrorHandler.class);
+        ProblemRestTemplateCustomizer customizer = new ProblemRestTemplateCustomizer(handler) {
+        };
+        RestTemplate restTemplate = new RestTemplate();
+        customizer.customize(restTemplate);
+
+        assertThat(restTemplate.getErrorHandler()).isSameAs(handler);
+    }
+
+}
