--- belgif-rest-problem-spring-boot-4-client-starter/src/test/java/io/github/belgif/rest/problem/spring/client/ClientProblemAutoConfigurationTest.java
@@ -1,22 +1,0 @@
-package io.github.belgif.rest.problem.spring.client;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.autoconfigure.AutoConfigurations;
-import org.springframework.boot.test.context.runner.ApplicationContextRunner;
-
-import io.github.belgif.rest.problem.spring.ProblemJackson3Configuration;
-
-class ClientProblemAutoConfigurationTest {
-
-    private final ApplicationContextRunner contextRunner = new ApplicationContextRunner()
-            .withConfiguration(AutoConfigurations.of(ClientProblemAutoConfiguration.class));
-
-    @Test
-    void autoConfiguration() {
-        contextRunner.run((context) -> {
-            assertThat(context).hasSingleBean(ProblemJackson3Configuration.class);
-        });
-    }
-}
