+++ belgif-rest-problem-spring-boot-4-server-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-io.github.belgif.rest.problem.spring.server.ServerProblemAutoConfiguration
+io.github.belgif.rest.problem.spring.boot.server.ServerProblemAutoConfiguration
