--- belgif-rest-problem-spring-boot-4-server-starter/src/test/java/io/github/belgif/rest/problem/spring/server/ServerProblemAutoConfigurationTest.java
@@ -1,23 +1,0 @@
-package io.github.belgif.rest.problem.spring.server;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.autoconfigure.AutoConfigurations;
-import org.springframework.boot.test.context.runner.WebApplicationContextRunner;
-
-import io.github.belgif.rest.problem.spring.ProblemJackson3Configuration;
-
-class ServerProblemAutoConfigurationTest {
-
-    private final WebApplicationContextRunner contextRunner = new WebApplicationContextRunner()
-            .withConfiguration(AutoConfigurations.of(ServerProblemAutoConfiguration.class));
-
-    @Test
-    void autoConfiguration() {
-        contextRunner.run((context) -> {
-            assertThat(context).hasSingleBean(ProblemJackson3Configuration.class);
-        });
-    }
-
-}
